chore: use pretty-quick
The `pretty-quick` package is basically a custom-built `lint-staged` for `prettier` only. It's 33x smaller than `lint-staged` and requires less configuration, but its also less flexible. If we want to use `eslint` on staged files only, we should stick with `lint-staged`, but we _could_ just use `eslint` in a pre-push hook instead.

Also, I don't know if `lint-staged` shares a lot of its dependencies with other packages we use, so it may not be as expensive as I think it is. Is there a way to track the number of transitive dependencies?